klarna/src/main/java/io/primer/android/klarna/implementation/payment/presentation/KlarnaPaymentDelegate.kt
package io.primer.android.klarna.implementation.payment.presentation

import io.primer.android.errors.data.exception.UnhandledPaymentPendingStateException
import io.primer.android.errors.domain.BaseErrorResolver
import io.primer.android.paymentmethods.common.data.model.PaymentMethodType
import io.primer.android.payments.core.create.domain.handler.PaymentMethodTokenHandler
import io.primer.android.domain.payments.create.model.Payment
import io.primer.android.payments.core.helpers.CheckoutErrorHandler
import io.primer.android.payments.core.helpers.CheckoutSuccessHandler
import io.primer.android.payments.core.helpers.PaymentMethodPaymentDelegate
import io.primer.android.payments.core.resume.domain.handler.PaymentResumeHandler

internal class KlarnaPaymentDelegate(
    paymentMethodTokenHandler: PaymentMethodTokenHandler,
    resumePaymentHandler: PaymentResumeHandler,
    successHandler: CheckoutSuccessHandler,
    errorHandler: CheckoutErrorHandler,
    baseErrorResolver: BaseErrorResolver
) : PaymentMethodPaymentDelegate(
    paymentMethodTokenHandler,
    resumePaymentHandler,
    successHandler,
    errorHandler,
    baseErrorResolver
) {

    override suspend fun handleNewClientToken(clientToken: String, payment: Payment?): Result<Unit> {
        return Result.failure(UnhandledPaymentPendingStateException(PaymentMethodType.KLARNA.name))
    }
}
